Add modules and move .vimrc from hidden
This was the first time I tried applying the repo on another
machine. It turns out that the bundle folders for the plugins I
thought would be there were actually missing. This is because
these folders were added via git. I had misunderstood this. I
thought the code would be copied. After looking a bit more, I
found a tutorial for this at:
http://vimcasts.org/episodes/synchronizing-plugins-with-git-submodules-and-pathogen/
I needed to add each sub repo as a submodule in .gitmodules and
then run the following commands:
git submodule init
git submodule update

Additionally, I found from this site that the convention is to
name the vimrc file in the repo without the . at the beginning.
So, I fixed that while I was at it.